requirements.txt
pytest
pytest-cov
pylint
pylint-report
Pillow
kivy==2.1.0
